bfd: Support PE associative COMDAT sections
Preserve PE/COFF Section Definition selection and parent metadata as an
explicit section graph.  Validate chains, make duplicate selection and
garbage collection follow the parent, and regenerate parent indices
after relocatable links and object copying.  Use the selection value as
the authoritative associative state and reject selector 5 at output
without a resolved output parent.  Handle full-width BigObj parent
indices and initialize the complete BigObj auxiliary record.

bfd/
        * Makefile.am (BFD_H_FILES): Add coff-bfd.c.
        * Makefile.in: Regenerate.
        * bfd-in2.h: Regenerate.
        * coff-bfd.c (bfd_coff_get_comdat_associative_parent): New.
        (bfd_coff_set_comdat_associative): New.
        * coff-bfd.h (coff_comdat_info): Record selection and parent.
        (coff_section_tdata): Add associative COMDAT graph state.
        (coff_section_data_associative): New.
        * coffcode.h (fill_comdat_hash, handle_COMDAT): Preserve Section
        Definition metadata.
        (coff_write_object_contents): Regenerate associative parent indices
        and reject unresolved associative metadata.
        (coff_bigobj_swap_aux_out): Handle and initialize 20-byte auxiliary
        records.
        * coffgen.c (coff_resolve_associative_comdats): New.
        (coff_kept_associative_section): New.
        (_bfd_coff_section_already_linked): Make children follow parent
        selection.
        (_bfd_coff_gc_mark): Traverse associative children.
        (coff_gc_sweep): Retain only legacy unassociated unwind sections.
        * cofflink.c (_bfd_coff_link_input_bfd): Rewrite associative parent
        indices.
        * libcoff-in.h (comdat_hash_entry): Preserve selection metadata.
        * libcoff.h: Regenerate.
        * peXXigen.c (_bfd_XX_bfd_copy_private_bfd_data_common): Rebuild an
        output-local association graph.
        (_bfd_XX_bfd_copy_private_section_data): Preserve serializable
        association metadata.

include/
        * coff/internal.h (internal_auxent): Widen x_associated.

+++ include/coff/internal.h
@@ -607,7 +607,7 @@ union internal_auxent
     uint16_t x_nreloc;		/* # relocation entries */
     uint16_t x_nlinno;		/* # line numbers */
     uint32_t x_checksum;	/* section COMDAT checksum for PE */
-    uint16_t x_associated;	/* COMDAT associated section index for PE */
+    uint32_t x_associated;      /* COMDAT associated section index for PE */
     uint8_t x_comdat;		/* COMDAT selection number for PE */
   } x_scn;
